// hrms-frontend/src/app/modules/employee/hosonhansu/model/ketQuaDaotaoCanhanUI.ts
export interface KetquadaotaocanhanUI {

	kequaDtId?: number;
  	nsId?: number;
  	donviId?: number;
  	nam?: number;
  	isRangbuoc?: boolean;
  	thoihanhetrangbuoc?: Date;
  	dtDanhmucDtId?: number;
  	trongnuoc?: boolean;
  	quocgiaId?: number;
	tennuoc?: string;
	tennuocList?: string[];
	ngaybd?: Date;
	ngaykt?: Date;
	kinhphi?: number;
	lydo?: string;
	ngayky?: Date;
	noidung?: string;
  	cmChinh?: boolean;
  	tungay?: Date;
	denngay?: Date;
	giohanhchinh?: boolean;
	hinhthucDtId?: number;
	hinhthucDt?: string;
	nsQdnoidungId?: number;
	nsQdndung?: any;
	fileName?: string;
	fileExtend?: string;
	file?: any;
	isChangeFileAttach?: boolean;
	soQd?: string;
	ngayQd?: Date;
	strNgayQd?: string;
	nguoiky?: string;
	chucvuky?: string;
	noidungQd?: string;
	coquantra?: boolean;
	tilytra?: number;
	sotientra?: number;
	sothangcamket?: number;
	nopketqua?: boolean;
	truongDtId?: number;
	tentruongDt?: string;
	trinhdoId?: number;
	trinhdo?: string;
	thoigiandtao?: string;
	nganhngheId?: number;
	nganhnge?: string;
	nnganhId?: number;
	ccSo?: string;
	ccNgaycap?: Date;
	ccNgayHH?: Date;
	sogiothucte?: number;
	dtngoaingu?: boolean;
	namtotnghiep?: number;
	xeploaiId?: number;
	xeploai?: string;
	noidaotao?: string;
	noidungdaotao?: string;
	ngoainguId?: number;
	ngoaingu?: string;
	trinhdongoainguId?: number;
	chungchi?: boolean;
	isdaihan?: boolean;
	isPass?: boolean;
	hocviId?: number;
	hocvi?: string;
	nsIDList?: number[];
	tenphongban?: string;
	tennhansu?: string;
	isCheck?: boolean;
	stt?: number;
	ketquadaotaocanhanUI?: KetquadaotaocanhanUI | null;
	saveList?: KetquadaotaocanhanUI[] | null;
	qtrinhdtNsGocId?: number;
	status?: number;
	filevbanId?: number;
	fileNameVB_CC?: string;
	fileExtendVB_CC?: string;
	fileVB_CC?: any;
	isChangeFileAttachVB_CC?: boolean;
	chungchiFileList?: any[];
	vanbangFileList?: any[];
	fileAttach?: any;
	fileQdName?: string;    
	quocgiaIdList?: any[];
}